fix: price's amount should use real type

+++ core/src/database/schemas/prices.sql
@@ -2,6 +2,6 @@ create table if not exists prices
 (
     datetime         datetime not null,
     commodity        varchar  not null,
-    amount           numeric  not null,
+    amount           REAL  not null,
     target_commodity varchar  not null
 );
